Change the look of the application by using bootstrap
