+++ .idea/.gitignore
@@ -1,3 +1,11 @@
 
 # Default ignored files
 /workspace.xml
+# Project exclude paths
+/.
+# Project exclude paths
+/.
+
+# Datasource local storage ignored files
+/dataSources.local.xml
+/dataSources/
