--- src/Table/Table.css
@@ -1,53 +1,0 @@
-.table {
-  background-color: white;
-  font-weight: bold;
-}
-
-.tableproperties {
-  background-color: white;
-  width: 40%;
-  display: inline-flex;
-}
-
-#pdf {
-  width: 50%;
-}
-
-.upload-btn-wrapper {
-  width: 50%;
-  position: relative;
-  overflow: hidden;
-  display: inline-block;
-}
-
-.upload-btn-wrapper input[type=file] {
-  font-size: 100px;
-  position: absolute;
-  left: 0;
-  top: 0;
-  opacity: 0;
-}
-
-.btn {
-  display: inline-block;
-  padding: 25px 17px;
-  font-size: 15px;
-  cursor: pointer;
-  text-align: center;
-  text-decoration: none;
-  outline: none;
-  color: #fff;
-  background-color: #4CAF50;
-  border: none;
-  border-radius: 15px;
-}
-
-.pdf_icon {
-  background-image: url(pdf_icon.png);
-  background-position: center left;
-  background-repeat: no-repeat;
-  padding-left: 80px;
-  height: 62px;
-  display: inline-block;
-  padding-top: 10px;
-}
