+++ src/components/Halls/Halls.js
@@ -3,7 +3,6 @@ import {BootstrapTable, TableHeaderColumn} from "react-bootstrap-table";
 import "react-bootstrap-table/dist/react-bootstrap-table-all.min.css";
 import styled from "styled-components";
 
-
 const Styles = styled.div`
 .btn {
     text-transform: uppercase;
@@ -11,32 +10,32 @@ const Styles = styled.div`
     }`
 ;
 
-class CandidatesTable extends Component {
+class HallsTable extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            candidates: []
+            halls: []
         };
     }
 
     componentDidMount() {
-        const url = "http://localhost:8080/candidates";
+        const url = "http://localhost:8080/halls";
         fetch(url, {
             method: "GET"
         })
             .then(response => response.json())
-            .then(candidates => {
+            .then(halls => {
                 this.setState({
-                    candidates: candidates
+                    halls: halls
                 });
             });
     }
 
     render() {
-        console.log(this.state.candidates);
+        console.log(this.state.halls);
         return (
             <Styles>
-                <BootstrapTable data={this.state.candidates}
+                <BootstrapTable data={this.state.halls}
                                 options={options}
                                 insertRow={true}
                                 deleteRow={true}
@@ -46,30 +45,15 @@ class CandidatesTable extends Component {
                                 selectRow={selectRowProp}
                                 cellEdit={cellEditProp}>
                     <TableHeaderColumn dataField='id' isKey hidden searchable={false}>ID</TableHeaderColumn>
-                    <TableHeaderColumn dataField='cnp' dataSort={true}>CNP</TableHeaderColumn>
-                    <TableHeaderColumn dataField='firstName'>Prenume</TableHeaderColumn>
-                    <TableHeaderColumn dataField='lastName' dataSort={true}>Nume</TableHeaderColumn>
-                    <TableHeaderColumn dataField='highSchool'>Liceu</TableHeaderColumn>
+                    <TableHeaderColumn dataField='name' dataSort={true}>Nume</TableHeaderColumn>
+                    <TableHeaderColumn dataField='utilizableSize' dataSort={true}>Nr Loc. Utilizate</TableHeaderColumn>
+                    <TableHeaderColumn dataField='size' dataSort={true}>Nr Locuri</TableHeaderColumn>
+                    <TableHeaderColumn>Candidati</TableHeaderColumn>
                 </BootstrapTable>
             </Styles>
         );
     }
 
-    updateCandidate(id, data) {
-        return fetch("http://localhost:8080/candidates/" + id, {
-            method: "PUT",
-            body: JSON.stringify(data),
-            headers: {
-                "Content-Type": "application/json"
-            }
-        })
-            .then(res => {
-                return res;
-            })
-            .catch(err => {
-                console.log(err);
-            });
-    }
 }
 
 const cellEditProp = {
@@ -87,6 +71,4 @@ const options = {
     saveText: 'my_save',
     closeText: 'my_close'
 };
-
-
-export default CandidatesTable;
+export default HallsTable;
