--- src/components/NavigationBar/NavigationBar.css
@@ -1,12 +1,0 @@
-.navbar {
-    background-color: #222;
-}
-
-.navbar-brand, .navbar-nav .nav-link {
-    color: #bbb;
-}
-
-.nav-link:hover {
-    color: white;
-}
-
